refactor : step3, step4 분리
