+++ src/main/java/step2/calculator/AvailableCalculation.java
