+++ src/main/java/step3/operation/Operation.java
@@ -1,11 +1,7 @@
 package step3.operation;
 
-import step3.domain.Car;
-import step3.utils.CarFactory;
-import step3.utils.FindWinners;
-import step3.view.InputView;
 import step3.view.ResultView;
-import java.util.List;
+
 import java.util.Random;
 
 public class Operation {
@@ -16,15 +12,6 @@ public class Operation {
         resultView.viewProgress(numOfAtp, numOfCars);
     }
 
-    public static void gameStartByName(int numOfAtp, String names) {
-        List<Car> cars = CarFactory.create(names);
-        InputView.inputTest(names, numOfAtp);
-
-        for(int i = 0; i < numOfAtp; i++) {
-            Operation.progressByName(cars);
-        }
-        FindWinners.findWinners(cars, numOfAtp);
-    }
 
     public void progress(int[] game, int[] result) {
         for(int w = 0 ; w < game.length; w++) {
@@ -49,24 +36,4 @@ public class Operation {
         return randomValue;
     }
 
-    public static void progressByName(List<Car> cars) {
-        StringBuilder sb = new StringBuilder();
-
-        for(int j = 0; j < cars.size(); j++) {
-            int position = cars.get(j).getPosition() + process();
-            cars.get(j).setPosition(position);
-            sb.append(ResultView.progressView(cars.get(j)));
-            sb.append(System.getProperty("line.separator"));
-        }
-        ResultView.view(sb.toString());
-    }
-
-    private static int process() {
-        Random random = new Random();
-        int randomNum = random.nextInt(10);
-        if(randomNum >= 4) {
-            return 1;
-        }
-        return 0;
-    }
 }
