+++ src/main/java/step4/domain/Car.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step4.domain;
 
 import java.util.Objects;
