+++ src/main/java/step4/utils/StringUtils.java
@@ -1,4 +1,4 @@
-package step3.utils;
+package step4.utils;
 
 public class StringUtils {
