+++ src/test/java/step3/RaceCarTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import step3.operation.Operation;
-import step3.utils.ValidationUtils;
 import step3.view.InputView;
 
 import static org.assertj.core.api.Assertions.assertThat;
