+++ src/test/java/step4/CarFactoryTest.java
@@ -1,9 +1,9 @@
-package step3;
+package step4;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.domain.Car;
-import step3.utils.CarFactory;
+import step4.domain.Car;
+import step4.utils.CarFactory;
 
 import java.util.List;
