+++ src/test/java/step4/UtilsTest.java
@@ -1,9 +1,9 @@
-package step3;
+package step4;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.utils.StringUtils;
-import step3.utils.ValidationUtils;
+import step4.utils.StringUtils;
+import step4.utils.ValidationUtils;
 
 import static org.assertj.core.api.Assertions.assertThat;
